task2/requirements.txt
aiohttp==3.4.4
ipython==6.5.0
motor==2.0.0
Pillow==5.2.0
